Machine and domicile shit

+++ VaskEnTid/VaskEnTidLib/Service/DomicileService.cs
@@ -31,14 +31,23 @@ namespace VaskEnTidLib.Service
         }
         public void RemoveUserByID(int userID, int domicileID)
         {
-            List<int> users = GetByUserID(userID).UserIDs;
-            foreach(int i in users) 
+            try 
             {
-                if (i == userID)
+                if (GetByUserID(userID) == null) { }
+                else
                 {
-                    _domicileRepo.RemoveUserByID(userID);
+                    List<int> users = GetByUserID(userID).UserIDs;
+                    foreach (int i in users)
+                    {
+                        if (i == userID)
+                        {
+                            _domicileRepo.RemoveUserByID(userID);
+                        }
+                    }
                 }
             }
+            catch { }
+            
             
         }
         public List<Domicile> GetAll()
